+++ core/src/com/fos/game/engine/ecs/systems/scripting/ScriptsUpdaterUtils.java
@@ -4,6 +4,6 @@ import com.fos.game.engine.ecs.components.base.ComponentType;
 
 public class ScriptsUpdaterUtils {
 
-    protected static final short SCRIPTED_ENTITY_BIT_MASK = ComponentType.SCRIPTS.bitMask;
+    protected static final int SCRIPTED_ENTITY_BIT_MASK = ComponentType.SCRIPTS.bitMask;
 
 }
